Renaming Phase III: BFColumn->FilterCoupler.
This anticipates a significant change in how this element works. It no
longer has table/attribute data. That will be handled by the requesting
module exclusively. It only keeps track of the applicable filter chain.
Modules may have multiple of these to represent data groupings that may
be filtered differently (e.g. nodes and links) which is independent
of which tables are there and may be applied to multiple in the same
way.